Temp Fix Dark Theme
