+++ app/src/main/java/com/jxareas/xpensor/data/local/converters/TimeConverter.kt
@@ -1,0 +1,18 @@
+package com.jxareas.xpensor.data.local.converters
+
+import androidx.room.TypeConverter
+import java.time.LocalTime
+
+class TimeConverter {
+
+    @TypeConverter
+    fun fromLocalTime(time: LocalTime): String {
+        return time.toString()
+    }
+
+    @TypeConverter
+    fun toLocalTime(data: String): LocalTime {
+        return LocalTime.parse(data)
+    }
+
+}
